ch refactor
