+++ .eslintrc.json
@@ -11,6 +11,7 @@
         "sourceType": "module"
     },
     "rules": {
-        "quotes": [2, "double", "avoid-escape"]
+        "quotes": [2, "double", "avoid-escape"],
+        
     }
 }
