--- src/database/models/bownerbusiness.js
@@ -1,10 +1,0 @@
-'use strict';
-module.exports = (sequelize, DataTypes) => {
-  const BownerBusiness = sequelize.define('BownerBusiness', {
-    businessOwnerId: DataTypes.INTEGER,
-    businessId: DataTypes.INTEGER,
-    status:DataTypes.INTEGER
-  }, {});
-
-  return BownerBusiness;
-};
