--- src/database/models/rolepermission.js
@@ -1,13 +1,0 @@
-'use strict';
-module.exports = (sequelize, DataTypes) => {
-  const rolepermission = sequelize.define('rolepermission', {
-
-    role_id:DataTypes.INTEGER,
-    permission_id: DataTypes.INTEGER,
-
-  }, {
-  timestamps:false,
-  });
- 
-  return rolepermission;
-};
