+++ src/resources/api/index.js
@@ -9,7 +9,7 @@ const apiVersion = process.env.API_VERSION;
 const baseUrl = `/api/${apiVersion}`;
 
 router.get("/api", (req, res) => {
-  res.status(200).json({ message: "Well Connected api" });
+    res.status(200).json({ message: "Well Connected api" });
 });
 
 router.use(baseUrl, api);
